Expand the ideas a bit further. Add some new ones.
Added the non-cooperation point as the defined reasonable way to react to those who oppose truth / reason. This comes immediately after the 'no harm' point to make it clear that the correct course of action when reacting to those against these beliefs is not to harm them.

I do believe that it needs to be defined how one can deal with those who are actively violent and harming others. Saying "harm those who harm others" just won't do as it is easy to justify harm to whoever you wish with such a notion. More thought needed.